wip: split FileEditor to modules
