+++ catalog/app/components/FileEditor/FileEditor.tsx
@@ -1,62 +1,16 @@
-import * as R from 'ramda'
 import * as React from 'react'
 import { useLocation } from 'react-router-dom'
-import * as M from '@material-ui/core'
 
-import * as AWS from 'utils/AWS'
 import AsyncResult from 'utils/AsyncResult'
 import parseSearch from 'utils/parseSearch'
 import type { S3HandleBase } from 'utils/s3paths'
 import * as PreviewUtils from 'components/Preview/loaders/utils'
 import PreviewDisplay from 'components/Preview/Display'
-import { detect as isMarkdown } from 'components/Preview/loaders/Markdown'
 
 import Skeleton from './Skeleton'
 import TextEditor from './TextEditor'
-import { Mode, EditorInputType } from './types'
-
-const cache: { [index in Mode]?: Promise<void> | 'fullfilled' } = {}
-const loadMode = (mode: Mode) => {
-  if (cache[mode] === 'fullfilled') return cache[mode]
-  if (cache[mode]) throw cache[mode]
-
-  cache[mode] = import(`brace/mode/${mode}`).then(() => {
-    cache[mode] = 'fullfilled'
-  })
-  throw cache[mode]
-}
-
-const isYaml = PreviewUtils.extIn(['.yaml', '.yml'])
-const typeYaml: EditorInputType = {
-  brace: 'yaml',
-}
-
-const typeMarkdown: EditorInputType = {
-  brace: 'markdown',
-}
-
-const typeNone: EditorInputType = {
-  brace: null,
-}
-
-const detect: (path: string) => EditorInputType = R.pipe(
-  PreviewUtils.stripCompression,
-  R.cond([
-    [isMarkdown, R.always(typeMarkdown)],
-    [isYaml, R.always(typeYaml)],
-    [R.T, R.always(typeNone)],
-  ]),
-)
-
-export function useWriteData({ bucket, key }: S3HandleBase) {
-  const s3 = AWS.S3.use()
-  return React.useCallback(
-    async (value) => {
-      await s3.putObject({ Bucket: bucket, Key: key, Body: value }).promise()
-    },
-    [bucket, key, s3],
-  )
-}
+import { EditorInputType } from './types'
+import { detect, loadMode, useWriteData } from './loader'
 
 interface EditorState {
   editing: boolean
@@ -93,18 +47,6 @@ export function useState(handle: S3HandleBase): EditorState {
       value,
     }),
     [editing, onCancel, onEdit, onSave, type, value],
-  )
-}
-
-interface AddFileButtonProps {
-  onClick: () => void
-}
-
-export function AddFileButton({ onClick }: AddFileButtonProps) {
-  return (
-    <M.Button variant="contained" color="primary" size="large" onClick={onClick}>
-      Create file
-    </M.Button>
   )
 }
